Move 2 SETUP operations from BASIS session to SETUP session
